soiliticalapi/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
numpy==1.24.4
scikit_learn==1.4.2
